requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# AI & NLP
openai==1.10.0
langchain==0.1.4
tiktoken==0.5.2
numpy==1.26.3

# Audio Processing & Transcription
openai-whisper==20231117
soundfile==0.12.1
pydub==0.25.1

# Integration - Zoom
zoom-python==1.0.0

# Integration - Google Meet & Calendar
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Integration - GitHub
PyGithub==2.2.0
pygithub==2.1.1

# Integration - Slack
slack-sdk==3.26.2
slack-bolt==1.18.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.3

# HTTP & Networking
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Task Queue & Background Jobs
celery==5.3.6
redis==5.0.1

# Data Validation & Parsing
python-dateutil==2.8.2
pytz==2024.1

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
